python/Docker-ServerPy/requirements.txt
Flask
pyyaml
kubernetes